requirements.txt
playwright
fastapi
uvicorn
requests
python-dotenv
openai
